修复了几乎所有的Warning，移除了所有PNG的ICC文件防止SRGB报错

+++ OOXX/OOXX-styleSheet.h
@@ -25,55 +25,55 @@ public:
 	SDL_Rect
 		//首页*********************
 		titlePage_title =		//首页标题
-		setRect(SCREEN_WIDTH*0.1, 100, SCREEN_WIDTH*0.8, SCREEN_HEIGHT*0.3),
+		setRect((int)(SCREEN_WIDTH*0.1), 100, (int)(SCREEN_WIDTH*0.8), (int)(SCREEN_HEIGHT*0.3)),
 
 		titlePage_btns[3] = {	//首页按钮集
-		setRect(SCREEN_WIDTH*0.5, SCREEN_HEIGHT*0.5, SCREEN_WIDTH*0.25, 60),
-		setRect(SCREEN_WIDTH*0.55, SCREEN_HEIGHT*0.65, SCREEN_WIDTH*0.25, 60),
-		setRect(SCREEN_WIDTH*0.6, SCREEN_HEIGHT*0.8, SCREEN_WIDTH*0.25, 60),
+		setRect((int)(SCREEN_WIDTH*0.5), (int)(SCREEN_HEIGHT*0.5), (int)(SCREEN_WIDTH*0.25), 60),
+		setRect((int)(SCREEN_WIDTH*0.55), (int)(SCREEN_HEIGHT*0.65), (int)(SCREEN_WIDTH*0.25), 60),
+		setRect((int)(SCREEN_WIDTH*0.6), (int)(SCREEN_HEIGHT*0.8), (int)(SCREEN_WIDTH*0.25), 60),
 	},
 
 
 		//游戏页********************
 		gamePage_chessBoard =	//游戏页棋盘
-		setRect(0, 0, SCREEN_WIDTH*0.75, SCREEN_HEIGHT),
+		setRect(0, 0, (int)(SCREEN_WIDTH*0.75), SCREEN_HEIGHT),
 
 		*gamePage_chessPoint = getChessPointRects(),	//81个棋子
 
 		gamePage_title =		//游戏页右侧小标题
-		setRect(SCREEN_WIDTH*0.77, 50, SCREEN_WIDTH*0.22, 100),
+		setRect((int)(SCREEN_WIDTH*0.77), 50, (int)(SCREEN_WIDTH*0.22), 100),
 
 		gamePage_cup =			//游戏页右侧下方咖啡杯
-		setRect(SCREEN_WIDTH*0.77, 500, SCREEN_WIDTH*0.22, 270),
+		setRect((int)(SCREEN_WIDTH*0.77), 500, (int)(SCREEN_WIDTH*0.22), 270),
 
 		*gamePage_midBox = getMidBoxRects(),
 
 		gamePage_nowTurn =		//游戏页右侧中间 "Now Turn" 字样区域
-		setRect(SCREEN_WIDTH*0.78, 250, SCREEN_WIDTH*0.21, 40),
+		setRect((int)(SCREEN_WIDTH*0.78), 250, (int)(SCREEN_WIDTH*0.21), 40),
 
 		gamePage_nowTurnIMG =	//游戏页右侧中间 当前回合棋子 图片
-		setRect(SCREEN_WIDTH*0.8, 305, 160, 160),
+		setRect((int)(SCREEN_WIDTH*0.8), 305, 160, 160),
 
 		//通知窗口******************
 		alert_alertPairCodeText =	//"请输入配对码"
-		setRect(SCREEN_WIDTH*0.18, SCREEN_HEIGHT*0.3, 400, 50),
+		setRect((int)(SCREEN_WIDTH*0.18), (int)(SCREEN_HEIGHT*0.3), 400, 50),
 
 		alert_wantExit =	//"你想要退出么？"
-		setRect(SCREEN_WIDTH*0.22, SCREEN_HEIGHT*0.32, 400, 60),
+		setRect((int)(SCREEN_WIDTH*0.22), (int)(SCREEN_HEIGHT*0.32), 400, 60),
 
 		alert_willLose =	//"你会失去当前状态"
-		setRect(SCREEN_WIDTH*0.18, SCREEN_HEIGHT*0.46, 500, 60),
+		setRect((int)(SCREEN_WIDTH*0.18), (int)(SCREEN_HEIGHT*0.46), 500, 60),
 
 		*alert_PairCode = getPairCodeRects(),	//四个配对码
 
 		alert_confirmBtn =						//通知窗口确定按钮
-		setRect(SCREEN_WIDTH*0.56, SCREEN_HEIGHT*0.61, 100, 70),
+		setRect((int)(SCREEN_WIDTH*0.56), (int)(SCREEN_HEIGHT*0.61), 100, 70),
 
 		alert_cancleBtn =						//通知窗口取消按钮
-		setRect(SCREEN_WIDTH*0.44, SCREEN_HEIGHT*0.61, 100, 70),
+		setRect((int)(SCREEN_WIDTH*0.44), (int)(SCREEN_HEIGHT*0.61), 100, 70),
 
 		alert_gameOver =							//红方胜利文本
-		setRect(SCREEN_WIDTH*0.18, SCREEN_HEIGHT*0.31, SCREEN_WIDTH *0.5, SCREEN_HEIGHT*0.25),
+		setRect((int)(SCREEN_WIDTH*0.18), (int)(SCREEN_HEIGHT*0.31), (int)(SCREEN_WIDTH *0.5), (int)(SCREEN_HEIGHT*0.25)),
 
 
 		//通用**********************
@@ -81,7 +81,7 @@ public:
 		setRect(SCREEN_WIDTH-60, SCREEN_HEIGHT-60, 50, 50),
 
 		common_alertWindow =	//通用的通知窗口
-		setRect(SCREEN_WIDTH*0.1, SCREEN_HEIGHT*0.2, SCREEN_WIDTH*0.8, SCREEN_HEIGHT*0.6);
+		setRect((int)(SCREEN_WIDTH*0.1), (int)(SCREEN_HEIGHT*0.2), (int)(SCREEN_WIDTH*0.8), (int)(SCREEN_HEIGHT*0.6));
 
 private:
 	SDL_Rect* getChessPointRects();
@@ -134,7 +134,7 @@ SDL_Rect* StyleSheet::getChessPointRects() {
 SDL_Rect* StyleSheet::getPairCodeRects() {
 	SDL_Rect* rects = new SDL_Rect[4];
 	for (int i = 0; i < 4; i++){
-		rects[i] = setRect(SCREEN_WIDTH*0.22+i*(30+SCREEN_WIDTH*0.08), SCREEN_HEIGHT*0.39, SCREEN_WIDTH*0.08, SCREEN_HEIGHT*0.16);
+		rects[i] = setRect((int)(SCREEN_WIDTH*0.22)+i*(30+ (int)(SCREEN_WIDTH*0.08)), (int)(SCREEN_HEIGHT*0.39), (int)(SCREEN_WIDTH*0.08), (int)(SCREEN_HEIGHT*0.16));
 	}
 	return rects;
 }
